+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         android:supportsRtl="false"
         android:theme="@android:style/Theme.Holo"
         android:usesCleartextTraffic="true">
-        <service android:name=".threads.weather.experimental.WeatherService" />
+        <service android:name=".threads.weather.WeatherService" />
 
         <activity
             android:name=".activities.OrariProcida2011Activity"
